This is a bit of a code cleanup.
I'm not sure if I got what you wanted to do with the Object check, but this is what it does now. Feel free to say if anything is wrong with it.

The code will check if you've passed it anything that isn't a jQuery/Zepto object. If you have, it'll do one of a few things. 

If it's a string, it'll search the form for an element that matches that selector. If it doesn't find anything, it'll error and stop there.

If you've passed it an object, it'll simply wrap it in jQuery/Zepto (If you pass a `HTMLElement` for example). 

If you haven't passed either of those things, it'll error there.

The rest of the code is pretty much the same, just formatted better.